cs50/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
cowsay==6.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
Joking==3.1.5
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pluggy==1.5.0
pyjoker==0.1.0
pyjokes==0.8.3
pytest==8.3.4
pytest-joke==0.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
tomli==2.2.1
tzdata==2024.2
urllib3==2.2.3
